fixed duplicate actions finally

+++ agents/canonicalise.py
@@ -552,7 +552,7 @@ class Canonicaliser(BaseAgent):
         obs_batch, nobs_batch, act_batch, done_batch, _, _, _, _, rew_batch, _ = sample
         dist, _, _ = self.policy.forward_with_embedding(obs_batch)
         if self.remove_duplicate_actions:
-            dist, act_batch = remove_duplicate_actions(dist, act_batch, self.env)
+            dist, act_batch = remove_duplicate_actions(dist, act_batch.to(torch.int32), self.env)
 
         val_batch = value_model(obs_batch).squeeze()
         next_val_batch = value_model(nobs_batch).squeeze()
